+++ engine/src/main/kotlin/engine/expressions/Expression.kt
@@ -269,13 +269,13 @@ open class Expression internal constructor(
 
     fun hasBracket() = decorators.isNotEmpty()
 
-    fun isPartialSum() = this is Sum && decorators.getOrNull(0) === Decorator.PartialBracket
-    fun isPartialProduct() = this is Product && decorators.getOrNull(0) === Decorator.PartialBracket
-
     /**
      * Any bracket other than Decorator.PartialBracket is a "visible bracket"
      */
-    fun hasVisibleBracket() = decorators.any { it != Decorator.PartialBracket }
+    fun hasVisibleBracket() = decorators.any { it != Decorator.PartialBracket && it != Decorator.MissingBracket }
+
+    fun isPartialSum() = this is Sum && decorators.getOrNull(0) === Decorator.PartialBracket
+    fun isPartialProduct() = this is Product && decorators.getOrNull(0) === Decorator.PartialBracket
 
     fun outerBracket() = decorators.lastOrNull()